Safari-611.1.10/LayoutTests/media/video-restricted-invisible-autoplay-allowed-if-audible-expected.txt


Test that "invisible autoplay not allowed restriction" does not pause audible media when scrolled out of view.

** setting video.src

EVENT(play)
EXPECTED (video.paused == 'false') OK
RUN(video.style.display = "none")
EXPECTED (video.paused == 'false') OK
RUN(video.style.removeProperty("display"))
RUN(video.style.visibility = "hidden")
EXPECTED (video.paused == 'false') OK
RUN(video.style.removeProperty("visibility"))
RUN(document.documentElement.style.height = window.innerHeight + 20 + video.offsetHeight)
RUN(window.scrollBy(0, 20 + video.offsetHeight))
EXPECTED (video.paused == 'false') OK
END OF TEST
